api/requirements.txt
fastapi==0.116.1
uvicorn==0.29.0
mlflow==2.22.0
pandas==2.2.3
python-dotenv==1.1.0
pydantic==2.11.5
boto3==1.38.36
scikit-learn==1.6.1
xgboost==3.0.2
numpy==2.2.6
scipy==1.15.3
psutil==7.0.0
setuptools==80.9.0
python-multipart==0.0.20